Enhance deabstraction to start producing GraphOperationInst in simple cases (#17463)
Enhance deabstraction to start producing GraphOperationInst in simple cases
when -tf-strict-deabstraction is enabled. Along with this, fix some minor
issues with GraphOperationInst handling metatypes, and add support for cloning
constants from the ConstExpr bump pointer to the ASTContext's when forming
instructions.

Lots of caveats remain:
- ConstExpr doesn't know about arrays.
- Partitioning doesn't know about GraphOperationInst.
- We don't handle input list operands.
- probably lots of other stuff to do.

The next steps after this is to wire up partitioning to act on these just
like they do other tensor ops.