[TypeChecker] Fix crash with Objective C keypaths.
Restore e0ce5fcbd905b1e615592ab4490b46f2abfeb5eb and
846f159a5cc2bb40de39f9531afa7a89f002fa86 without the change to have
the constraint generator bail out on failure as that seemed to cause a
verifier failure in CoreStore in the compatibility suite.

We have a crash reported but do not have a test case and my attempt at
creating one was unsuccessful. It seems like the issue is that we have
an invalid decl that we attempt to access the interface type on,
though, so hopefully this fixes the issue.

I also tweaked the constraint generator to bail out when
checkObjCKeyPathExpr fails.

Potentially fixes: rdar://problem/40955755