Added an overloaded version of toHost() and toAccelerator() that takes a tensor shape, to be used for XLA/TPU compilation (#17620)
Added an overloaded version of toHost() and toAccelerator() that takes a tensor
shape, to be used for XLA/TPU compilation.

When lowering to TPU infeed enqueue op, make the shapes attr optional. If specified, it's only used for error checking, so the enqueue ops in general know about the tensor shape of their input tensor. Note the shapes attr for infeed/output dequeue ops is still required.