Remove the "shrink" part of the solver.
At one point this phase was providing a lot of value in reducing the
number of expressions we ultimately decided were too complex, but at
this point it doesn't seem to be helping very often.

Based on experimenting with some of our slow test cases, in several
cases it actually increases the amount of work we do before we decide
an expression is too complex, sometimes making it take 2-3x longer.

I will open a bug to track restoring the two expressions that we are
now "too complex" on, so that the tests can be restored once we fix
some of the underlying problems.