Added partial support for SwitchEnum, which works when the num has exactly 2 cases, such as {some, none} in a swift optional.
A later PR will extend the support for N cases.

The design is for host to send the switch case id (a int64 value) to accelerator
in each of the cases, and for accelerator to recv the case id, and use a
sequence of cond_br instructions to dispatch to the right basic block.

To illustrate this design with pseudo-code in SIL, host does:

bb0:
  ...
  switch_enum %X, bb1, bb2, ..., bbN

bb1:
  sendToAccelerator(0) // 0 is the case id here
  <other host code>
  br bb_merge

<same for bb2 through bbN, sending a different case_id each>

bb_merge:
  ...

Accelerator side:
bb0:
  ...
  %case_id = recvFromHost
  if %case_id == 0 goto bb1
  else if %case_id == 1 goto bb2
  else ...

When TF backend supports switch_case in the future, we can lower to switch_case
instead of a sequence of if's / cond_br's.