[Type checker] Make “override” computation not dependent on Objective-C interop.
For reasons lost to history and not interesting enough to uncover, the
“override” computation performed a check of the overriding declaration’s
Objective-C selector against the potentially-overridden declaration’s
Objective-C selector. This introduced a cycle dependency between
the “is exposed to Objective-C” computation and the “overridden declarations”
computation, because @objc is inherited from an overridden declaration.

We’ll detect the same errors via other, existing diagnostics, so prune
out the Objective-C selector checks from the override path.