[Type Checker] Only resolveOverriddenDecl() for decls with ‘override’.
resolveOverriddenDecl() is meant to be used to provide access to the
overridden declarations, e.g., from another source file. Only resolve
when the declaration has an explicit ‘override’ modifier or is an
initializer (for which there are cases where ‘override’ is not
required).

This does not yet reduce override checking across source files,
because checkOverrides() is still called during
validateDeclForNameLookup(), and we’re still doing too much work to
compute overrides.