For a tfop builtin, take its parameters at +0 instead of +1 (#17601)
For a tfop builtin, take its parameters at +0 instead of +1, similar to the
recent swift calling convention change from @owned to @guaranteed.

This reduces the # of retain/release insts generated, and also allows us to
remove tfop insts from SIL code (e.g. when cleaning up the host function at the
end of the partitioning pass) without having to worry about their effects on the
refcount rebiases of the operands.

The next step is to apply the same change to the new graph op inst.

Also re-enabled sil-ownership when building the stdlib TF module.