Fix debug scope verification.
Debug scope verification was enabled by default in
https://github.com/apple/swift/commit/f917c9355764dc6d9308ccf2801776c62e032612.

This caused verification errors as a result of deabstraction inlining.
It's stated that `verifyDebugScopeHoles` should not be run when inlining is
performed because scopes are moved around.

The simplest patch is to mark inlined functions as being being optimized.
A more robust solution would be to check whether inlining has been performed
during verification.

Addresses SR-8114.