Disable 'self' parameter for #gradient and #valueAndGradient.
Disable differentiation with respect to 'self' for #gradient and #valueAndGradient.
'self' in the parameter list is ad-hoc and out of place.

The same "diff w.r.t. self" functionality can be generalized and represented
more naturally as dataflow AD, using a differential operator on an output
variable w.r.t input variables.