[dev-scripts] Add a new tool called scurve_printer to print scurve data produced by csvcolumn_to_scurve.
This used matplotlib internally to plot the values. Now one can produce scurves
without leaving the command line.

I tried to get logarithmic scale to work, but I was unable to do so in a trivial
way. That is something for a different hacking session. I need to move on.