ConstExpr improvements for strings and moving towards arrays.
Improve ConstExpr processing for strings to handle ones with literals correctly.
Start treating Builtin.RawPointer as addresses, which will be necessary for array
literals.

Refactor the code that analyzes calls to _allocateUninitializedArray and move it into
ConstExprEvaluator (not the best place perhaps, but better than what we had before).

Add a testcase showing that we handle empty and non-empty strings now.