Merge Address propagation and handling back into SymbolicValue. (#17823)
This moves us towards a model where SymbolicValue can track addresses.
I avoided this originally because they are nearly distinct concepts in
SIL, but such an attempt was doomed to fail: SIL has operations that
convert addresses to Builtin.RawPointer and ptrtoint to convert it to
integers.  These all get used with strings and arrays, so we need to
model them as SymbolicValues.

The heap model is also incorrect, but I'll address that in the next
patch.