Fixed a bug in building `GraphFunctionBody::outputs`, where the lowering of nested conditional SESE regions did not set that vector field.
The solution is to track the reads and writes of `outputs` more rigorously via
an enum field `GraphFunctionBody::OutputSetter`. Before this patch, we check
whether `outputs` have been set via `outputs.empty()`, but some graph functions
could have an empty output, so this is not sufficient. Also, the check was not
done in all contexts where `outputs` are written to.

This change is needed to unblock https://github.com/apple/swift/pull/17698, and
is split out of that patch.