Fixed a bug in TFFunctionPartition::markBlock() when handling a loop. (#17785)
Fixed a bug in TFFunctionPartition::markBlock() when handling a loop.

An example CFG: bb0 -> bb1 -> {bb2, bb3}; bb2 -> bb1
Here bb1 is the loop header, bb2 the loop body.

In this case, when we run markBlock() over bb1, we first find bb2,
post-dominated by bb1. We then walk bb2 (pushing it to worklist), where we find
bb1 as its predecessor -- in that case we must mark the term inst of bb1 (a
cond_br), so that the loop count variable gets promoted to tensor.

The comment blocks in the code and added test cases provide more context.

Also renamed `visited` to `addedToList` to better reflect its semantics, and
moved around sanity check code on `numTensorSuccs` to make it clear that the
core marking logic does not depend on it.