Fixed a bug in createHostSend(), where the scalar to send may not be produced by a struct_extract inst (#17890)
Fixed a bug in createHostSend(), where the scalar to send may not be produced by
a struct_extract inst. Instead of patterning matching on inst type, we now
pattern match on the data type.

The test case is now failing with a different error related to loop
canonicalization (https://bugs.swift.org/browse/SR-7765), so it's commented out.