[4.2] SILGen: When transforming to AnyHashable materialize the value in a temporary
emitAnyHashable expects an address value.

Cherry-pick from #17928.

Explanation: When transforming values to AnyHashable we sometimes (e.g
tuples of strings) omit converting the value to an address type by
storing it to the stack leading to a compiler crash.
The fix is to store the value to the stack if it is not already
available as an in memory value.

Scope: Affects only conversions to AnyHashable. The code seems to stem
from 2016.

Risk: Low. Code is added that checks that the value expected by
emitAnyHashable is already in memory if it is not we store it to memory.

Testing: A Swift CI test was added.

rdar://40583597